+++ app/Http/Controllers/SurveyController.php
@@ -32,17 +32,19 @@ class SurveyController extends Controller
     }
 
     public function show() {
-        $surveys = Survey::where('visible', 1)->paginate(10);
+        $surveys = Survey::where('visible', 1)->orderBy('created_at', 'DESC')->paginate(10);
         return view('pages.survey', compact('surveys', 'users'));
     }
 
     public function makeVisible(Survey $survey) {
+        $this->authorize('view', $survey);
         $survey->visible = 1;
         $survey->save();
         return redirect('/');
     }
 
     public function makeVisiblePrivately(Survey $survey) {
+        $this->authorize('view', $survey);
         $survey->visible = 2;
         $survey->save();
         return redirect('/');
@@ -52,12 +54,13 @@ class SurveyController extends Controller
         return view('survey.takesurvey', compact('survey'));
     }
 
-    public function showAnalytics() {
-        $survey = Survey::all();
+    public function showAnalytics(Survey $survey) {
+        $this->authorize('view', $survey);
         return view('survey.show-analytics', compact('survey'));
     }
 
     public function delete(Survey $survey) {
+        $this->authorize('view', $survey);
         foreach ($survey->questions as $question) {
             foreach($question->answers as $answer) {
                 $answer->delete();
@@ -65,6 +68,24 @@ class SurveyController extends Controller
             $question->delete();
         }
         $survey->delete();
+        return redirect('/');
+    }
+
+    public function editShow(Survey $survey) {
+        $this->authorize('view', $survey);
+        return view('survey.edit', compact('survey'));
+    }
+
+    public function edit(Survey $survey) {
+        $this->authorize('view', $survey);
+        $data = request()->validate([
+            'title' => 'required|max:100',
+            'description' => 'required|max:255',
+        ]);
+        $survey->title = $data['title'];
+        $survey->description = $data['description'];
+        $survey->save();
+
         return redirect('/');
     }
